+++ packages/fighting-design/drawer/src/props.ts
@@ -2,14 +2,13 @@ import {
   setBooleanProp,
   setObjectProp,
   setStringProp,
-  setStringNumberProp,
   setNumberProp,
   setFunctionProp
 } from '../../_utils'
 import type { ExtractPropTypes } from 'vue'
 import type { DrawerDirection } from './interface'
 import type { FightingIcon } from '../../_interface'
-import type { PopupCallback } from '../../popup'
+import type { DialogCallback } from '../../dialog'
 
 export const Props = {
   /**绑定值，控制是否展示 */
@@ -20,12 +19,12 @@ export const Props = {
   direction: setStringProp<DrawerDirection>('right', (val: DrawerDirection): boolean => {
     return (['left', 'top', 'right', 'bottom'] as const).includes(val)
   }),
-  /** 大小尺寸 */
-  size: setStringNumberProp('30%'),
   /** 标题文字内容 */
   title: setStringProp(),
   /** 是否展示遮罩层 */
   showMask: setBooleanProp(true),
+  /** 是否展示头部 */
+  showHeader: setBooleanProp(true),
   /** 点击遮罩层是否可关闭 */
   maskClose: setBooleanProp(true),
   /** 是否模糊遮罩层 */
@@ -41,13 +40,13 @@ export const Props = {
    */
   zIndex: setNumberProp(),
   /** 打开动画开始的回调 */
-  onOpen: setFunctionProp<PopupCallback>(),
+  onOpen: setFunctionProp<DialogCallback>(),
   /** 打开动画结束的回调 */
-  onOpenEnd: setFunctionProp<PopupCallback>(),
+  onOpenEnd: setFunctionProp<DialogCallback>(),
   /** 关闭动画开始的回调 */
-  onClose: setFunctionProp<PopupCallback>(),
+  onClose: setFunctionProp<DialogCallback>(),
   /** 关闭动画结束的回调 */
-  onCloseEnd: setFunctionProp<PopupCallback>()
+  onCloseEnd: setFunctionProp<DialogCallback>()
 } as const
 
 /** drawer 组件 props 类型 */