+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 - 优化 `f-text` 组件样式细节
 - 废除 `f-dialog` height 配置项
+- 废除 `f-drawer` size 配置项
 - 废除 `f-popup` 组件
 - 重构 `f-dialog` `f-drawer` 组件
+- `f-drawer` `f-dialog` 组件新增 `show-header` 配置项
 
 ## 0.41.0 (2023-06-04)